refactor: restructure app folder structure
