+++ app/dependencies/settings.py
@@ -1,7 +1,7 @@
 import os
 from functools import lru_cache
 
-from .config import Settings
+from app.core.config import Settings
 
 @lru_cache
 def get_settings():
